+++ client/shim.js
@@ -1,5 +1,5 @@
 /**
- * core-js 2.1.5
+ * core-js 2.2.0
  * https://github.com/zloirock/core-js
  * License: http://rock.mit-license.org
  * © 2016 Denis Pushkarev
@@ -178,10 +178,9 @@
 	__webpack_require__(223);
 	__webpack_require__(224);
 	__webpack_require__(225);
-	__webpack_require__(228);
-	__webpack_require__(229);
-	__webpack_require__(232);
-	__webpack_require__(233);
+	__webpack_require__(227);
+	__webpack_require__(230);
+	__webpack_require__(231);
 	__webpack_require__(234);
 	__webpack_require__(235);
 	__webpack_require__(236);
@@ -191,21 +190,21 @@
 	__webpack_require__(240);
 	__webpack_require__(241);
 	__webpack_require__(242);
+	__webpack_require__(243);
 	__webpack_require__(244);
-	__webpack_require__(245);
 	__webpack_require__(246);
 	__webpack_require__(247);
 	__webpack_require__(248);
+	__webpack_require__(249);
 	__webpack_require__(250);
 	__webpack_require__(251);
+	__webpack_require__(253);
 	__webpack_require__(254);
-	__webpack_require__(255);
 	__webpack_require__(256);
 	__webpack_require__(257);
 	__webpack_require__(258);
 	__webpack_require__(259);
-	__webpack_require__(260);
-	__webpack_require__(261);
+	__webpack_require__(262);
 	__webpack_require__(263);
 	__webpack_require__(264);
 	__webpack_require__(265);
@@ -213,10 +212,17 @@
 	__webpack_require__(267);
 	__webpack_require__(268);
 	__webpack_require__(269);
-	__webpack_require__(270);
 	__webpack_require__(271);
 	__webpack_require__(272);
-	module.exports = __webpack_require__(273);
+	__webpack_require__(273);
+	__webpack_require__(274);
+	__webpack_require__(275);
+	__webpack_require__(276);
+	__webpack_require__(277);
+	__webpack_require__(278);
+	__webpack_require__(279);
+	__webpack_require__(280);
+	module.exports = __webpack_require__(281);
 
 
 /***/ },
@@ -255,11 +261,13 @@
 	  , $JSON          = global.JSON
 	  , _stringify     = $JSON && $JSON.stringify
 	  , setter         = false
+	  , PROTOTYPE      = 'prototype'
 	  , HIDDEN         = wks('_hidden')
+	  , TO_PRIMITIVE   = wks('toPrimitive')
 	  , isEnum         = {}.propertyIsEnumerable
 	  , SymbolRegistry = shared('symbol-registry')
 	  , AllSymbols     = shared('symbols')
-	  , ObjectProto    = Object.prototype
+	  , ObjectProto    = Object[PROTOTYPE]
 	  , USE_NATIVE     = typeof $Symbol == 'function'
 	  , QObject        = global.QObject;
 
@@ -276,7 +284,7 @@
 	} : dP;
 
 	var wrap = function(tag){
-	  var sym = AllSymbols[tag] = _create($Symbol.prototype);
+	  var sym = AllSymbols[tag] = _create($Symbol[PROTOTYPE]);
 	  sym._k = tag;
 	  DESCRIPTORS && setter && setSymbolDesc(ObjectProto, tag, {
 	    configurable: true,
@@ -288,8 +296,10 @@
 	  return sym;
 	};
 
-	var isSymbol = function(it){
+	var isSymbol = USE_NATIVE && typeof $Symbol.iterator == 'symbol' ? function(it){
 	  return typeof it == 'symbol';
+	} : function(it){
+	  return it instanceof $Symbol;
 	};
 
 	var $defineProperty = function defineProperty(it, key, D){
@@ -369,16 +379,12 @@
 	// 19.4.1.1 Symbol([description])
 	if(!USE_NATIVE){
 	  $Symbol = function Symbol(){
-	    if(isSymbol(this))throw TypeError('Symbol is not a constructor');
+	    if(this instanceof $Symbol)throw TypeError('Symbol is not a constructor!');
 	    return wrap(uid(arguments.length > 0 ? arguments[0] : undefined));
 	  };
-	  redefine($Symbol.prototype, 'toString', function toString(){
+	  redefine($Symbol[PROTOTYPE], 'toString', function toString(){
 	    return this._k;
 	  });
-
-	  isSymbol = function(it){
-	    return it instanceof $Symbol;
-	  };
 
 	  $GOPD.f = $getOwnPropertyDescriptor;
 	  $DP.f   = $defineProperty;
@@ -414,7 +420,7 @@
 	};
 
 	// Don't use setters in Qt Script, https://github.com/zloirock/core-js/issues/173
-	if(!QObject || !QObject.prototype || !QObject.prototype.findChild)setter = true;
+	if(!QObject || !QObject[PROTOTYPE] || !QObject[PROTOTYPE].findChild)setter = true;
 
 	$export($export.S + $export.F * !USE_NATIVE, 'Symbol', {
 	  // 19.4.2.1 Symbol.for(key)
@@ -425,7 +431,8 @@
 	  },
 	  // 19.4.2.5 Symbol.keyFor(sym)
 	  keyFor: function keyFor(key){
-	    return keyOf(SymbolRegistry, key);
+	    if(isSymbol(key))return keyOf(SymbolRegistry, key);
+	    throw TypeError(key + ' is not a symbol!');
 	  },
 	  useSetter: function(){ setter = true; },
 	  useSimple: function(){ setter = false; }
@@ -449,6 +456,8 @@
 	// 24.3.2 JSON.stringify(value [, replacer [, space]])
 	$JSON && $export($export.S + $export.F * (!USE_NATIVE || BUGGY_JSON), 'JSON', {stringify: $stringify});
 
+	// 19.4.3.4 Symbol.prototype[@@toPrimitive](hint)
+	$Symbol[PROTOTYPE][TO_PRIMITIVE] || __webpack_require__(8)($Symbol[PROTOTYPE], TO_PRIMITIVE, $Symbol[PROTOTYPE].valueOf);
 	// 19.4.3.5 Symbol.prototype[@@toStringTag]
 	setToStringTag($Symbol, 'Symbol');
 	// 20.2.1.9 Math[@@toStringTag]
@@ -469,7 +478,7 @@
 /* 3 */
 /***/ function(module, exports) {
 
-	var core = module.exports = {version: '2.1.5'};
+	var core = module.exports = {version: '2.2.0'};
 	if(typeof __e == 'number')__e = core; // eslint-disable-line no-undef
 
 /***/ },
@@ -5162,10 +5171,33 @@
 /* 225 */
 /***/ function(module, exports, __webpack_require__) {
 
+	var TO_PRIMITIVE = __webpack_require__(23)('toPrimitive')
+	  , proto        = Date.prototype;
+
+	if(!(TO_PRIMITIVE in proto))__webpack_require__(8)(proto, TO_PRIMITIVE, __webpack_require__(226));
+
+/***/ },
+/* 226 */
+/***/ function(module, exports, __webpack_require__) {
+
+	'use strict';
+	var anObject    = __webpack_require__(10)
+	  , toPrimitive = __webpack_require__(14)
+	  , NUMBER      = 'number';
+
+	module.exports = function(hint){
+	  if(hint !== 'string' && hint !== NUMBER && hint !== 'default')throw TypeError('Incorrect hint');
+	  return toPrimitive(anObject(this), hint != NUMBER);
+	};
+
+/***/ },
+/* 227 */
+/***/ function(module, exports, __webpack_require__) {
+
 	'use strict';
 	var $export      = __webpack_require__(7)
-	  , $typed       = __webpack_require__(226)
-	  , buffer       = __webpack_require__(227)
+	  , $typed       = __webpack_require__(228)
+	  , buffer       = __webpack_require__(229)
 	  , anObject     = __webpack_require__(10)
 	  , toIndex      = __webpack_require__(34)
 	  , toLength     = __webpack_require__(32)
@@ -5211,7 +5243,7 @@
 	__webpack_require__(183)(ARRAY_BUFFER);
 
 /***/ },
-/* 226 */
+/* 228 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var global = __webpack_require__(2)
@@ -5242,14 +5274,14 @@
 	};
 
 /***/ },
-/* 227 */
+/* 229 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
 	var global         = __webpack_require__(2)
 	  , DESCRIPTORS    = __webpack_require__(5)
 	  , LIBRARY        = __webpack_require__(47)
-	  , $typed         = __webpack_require__(226)
+	  , $typed         = __webpack_require__(228)
 	  , hide           = __webpack_require__(8)
 	  , redefineAll    = __webpack_require__(198)
 	  , fails          = __webpack_require__(6)
@@ -5522,26 +5554,26 @@
 	exports[DATA_VIEW] = $DataView;
 
 /***/ },
-/* 228 */
+/* 230 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var $export = __webpack_require__(7);
-	$export($export.G + $export.W + $export.F * !__webpack_require__(226).ABV, {
-	  DataView: __webpack_require__(227).DataView
+	$export($export.G + $export.W + $export.F * !__webpack_require__(228).ABV, {
+	  DataView: __webpack_require__(229).DataView
 	});
 
 /***/ },
-/* 229 */
+/* 231 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Int8', 1, function(init){
+	__webpack_require__(232)('Int8', 1, function(init){
 	  return function Int8Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
 
 /***/ },
-/* 230 */
+/* 232 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
@@ -5550,8 +5582,8 @@
 	    , global              = __webpack_require__(2)
 	    , fails               = __webpack_require__(6)
 	    , $export             = __webpack_require__(7)
-	    , $typed              = __webpack_require__(226)
-	    , $buffer             = __webpack_require__(227)
+	    , $typed              = __webpack_require__(228)
+	    , $buffer             = __webpack_require__(229)
 	    , ctx                 = __webpack_require__(18)
 	    , anInstance          = __webpack_require__(82)
 	    , propertyDesc        = __webpack_require__(15)
@@ -5571,7 +5603,7 @@
 	    , create              = __webpack_require__(41)
 	    , getPrototypeOf      = __webpack_require__(55)
 	    , gOPN                = __webpack_require__(45).f
-	    , isIterable          = __webpack_require__(231)
+	    , isIterable          = __webpack_require__(233)
 	    , getIterFn           = __webpack_require__(154)
 	    , uid                 = __webpack_require__(17)
 	    , wks                 = __webpack_require__(23)
@@ -5769,9 +5801,6 @@
 	        that[length]  = value;
 	      } return that;
 	    },
-	    slice: function slice(start, end){
-	      return speciesFromList(this, arraySlice.call(validate(this), start, end));
-	    },
 	    some: function some(callbackfn /*, thisArg */){
 	      return arraySome(validate(this), callbackfn, arguments.length > 1 ? arguments[1] : undefined);
 	    },
@@ -5788,6 +5817,10 @@
 	        toLength((end === undefined ? length : toIndex(end, length)) - $begin)
 	      );
 	    }
+	  };
+
+	  var $slice = function slice(start, end){
+	    return speciesFromList(this, arraySlice.call(validate(this), start, end));
 	  };
 
 	  var $set = function set(arrayLike /*, offset */){
@@ -5861,6 +5894,7 @@
 	  redefineAll($TypedArrayPrototype$, $iterators);
 	  hide($TypedArrayPrototype$, ITERATOR, $iterators.values);
 	  redefineAll($TypedArrayPrototype$, {
+	    slice:          $slice,
 	    set:            $set,
 	    constructor:    function(){ /* noop */ },
 	    toString:       arrayToString,
@@ -6001,11 +6035,17 @@
 
 	    $export($export.P, NAME, proto);
 
+	    setSpecies(NAME);
+
 	    $export($export.P + $export.F * FORCED_SET, NAME, {set: $set});
 
 	    $export($export.P + $export.F * !CORRECT_ITER_NAME, NAME, $iterators);
 
 	    $export($export.P + $export.F * (TypedArrayPrototype.toString != arrayToString), NAME, {toString: arrayToString});
+
+	    $export($export.P + $export.F * fails(function(){
+	      new TypedArray(1).slice();
+	    }), NAME, {slice: $slice});
 
 	    $export($export.P + $export.F * (fails(function(){
 	      return [1, 2].toLocaleString() != new TypedArray([1, 2]).toLocaleString()
@@ -6015,13 +6055,11 @@
 
 	    Iterators[NAME] = CORRECT_ITER_NAME ? $nativeIterator : $iterator;
 	    if(!LIBRARY && !CORRECT_ITER_NAME)hide(TypedArrayPrototype, ITERATOR, $iterator);
-
-	    setSpecies(NAME);
 	  };
 	} else module.exports = function(){ /* empty */ };
 
 /***/ },
-/* 231 */
+/* 233 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var classof   = __webpack_require__(71)
@@ -6035,31 +6073,11 @@
 	};
 
 /***/ },
-/* 232 */
-/***/ function(module, exports, __webpack_require__) {
-
-	__webpack_require__(230)('Uint8', 1, function(init){
-	  return function Uint8Array(data, byteOffset, length){
-	    return init(this, data, byteOffset, length);
-	  };
-	});
-
-/***/ },
-/* 233 */
-/***/ function(module, exports, __webpack_require__) {
-
-	__webpack_require__(230)('Uint8', 1, function(init){
-	  return function Uint8ClampedArray(data, byteOffset, length){
-	    return init(this, data, byteOffset, length);
-	  };
-	}, true);
-
-/***/ },
 /* 234 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Int16', 2, function(init){
-	  return function Int16Array(data, byteOffset, length){
+	__webpack_require__(232)('Uint8', 1, function(init){
+	  return function Uint8Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
@@ -6068,18 +6086,18 @@
 /* 235 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Uint16', 2, function(init){
-	  return function Uint16Array(data, byteOffset, length){
+	__webpack_require__(232)('Uint8', 1, function(init){
+	  return function Uint8ClampedArray(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
-	});
+	}, true);
 
 /***/ },
 /* 236 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Int32', 4, function(init){
-	  return function Int32Array(data, byteOffset, length){
+	__webpack_require__(232)('Int16', 2, function(init){
+	  return function Int16Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
@@ -6088,8 +6106,8 @@
 /* 237 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Uint32', 4, function(init){
-	  return function Uint32Array(data, byteOffset, length){
+	__webpack_require__(232)('Uint16', 2, function(init){
+	  return function Uint16Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
@@ -6098,8 +6116,8 @@
 /* 238 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Float32', 4, function(init){
-	  return function Float32Array(data, byteOffset, length){
+	__webpack_require__(232)('Int32', 4, function(init){
+	  return function Int32Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
@@ -6108,8 +6126,8 @@
 /* 239 */
 /***/ function(module, exports, __webpack_require__) {
 
-	__webpack_require__(230)('Float64', 8, function(init){
-	  return function Float64Array(data, byteOffset, length){
+	__webpack_require__(232)('Uint32', 4, function(init){
+	  return function Uint32Array(data, byteOffset, length){
 	    return init(this, data, byteOffset, length);
 	  };
 	});
@@ -6118,12 +6136,32 @@
 /* 240 */
 /***/ function(module, exports, __webpack_require__) {
 
+	__webpack_require__(232)('Float32', 4, function(init){
+	  return function Float32Array(data, byteOffset, length){
+	    return init(this, data, byteOffset, length);
+	  };
+	});
+
+/***/ },
+/* 241 */
+/***/ function(module, exports, __webpack_require__) {
+
+	__webpack_require__(232)('Float64', 8, function(init){
+	  return function Float64Array(data, byteOffset, length){
+	    return init(this, data, byteOffset, length);
+	  };
+	});
+
+/***/ },
+/* 242 */
+/***/ function(module, exports, __webpack_require__) {
+
 	'use strict';
+	// https://github.com/tc39/Array.prototype.includes
 	var $export   = __webpack_require__(7)
 	  , $includes = __webpack_require__(31)(true);
 
 	$export($export.P, 'Array', {
-	  // https://github.com/domenic/Array.prototype.includes
 	  includes: function includes(el /*, fromIndex = 0 */){
 	    return $includes(this, el, arguments.length > 1 ? arguments[1] : undefined);
 	  }
@@ -6132,7 +6170,7 @@
 	__webpack_require__(175)('includes');
 
 /***/ },
-/* 241 */
+/* 243 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
@@ -6147,12 +6185,13 @@
 	});
 
 /***/ },
-/* 242 */
+/* 244 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
+	// https://github.com/tc39/proposal-string-pad-start-end
 	var $export = __webpack_require__(7)
-	  , $pad    = __webpack_require__(243);
+	  , $pad    = __webpack_require__(245);
 
 	$export($export.P, 'String', {
 	  padStart: function padStart(maxLength /*, fillString = ' ' */){
@@ -6161,7 +6200,7 @@
 	});
 
 /***/ },
-/* 243 */
+/* 245 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/tc39/proposal-string-pad-start-end
@@ -6184,12 +6223,13 @@
 
 
 /***/ },
-/* 244 */
+/* 246 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
+	// https://github.com/tc39/proposal-string-pad-start-end
 	var $export = __webpack_require__(7)
-	  , $pad    = __webpack_require__(243);
+	  , $pad    = __webpack_require__(245);
 
 	$export($export.P, 'String', {
 	  padEnd: function padEnd(maxLength /*, fillString = ' ' */){
@@ -6198,7 +6238,7 @@
 	});
 
 /***/ },
-/* 245 */
+/* 247 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
@@ -6210,7 +6250,7 @@
 	}, 'trimStart');
 
 /***/ },
-/* 246 */
+/* 248 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
@@ -6222,10 +6262,45 @@
 	}, 'trimEnd');
 
 /***/ },
-/* 247 */
+/* 249 */
 /***/ function(module, exports, __webpack_require__) {
 
-	// https://gist.github.com/WebReflection/9353781
+	'use strict';
+	// https://tc39.github.io/String.prototype.matchAll/
+	var $export     = __webpack_require__(7)
+	  , defined     = __webpack_require__(30)
+	  , toLength    = __webpack_require__(32)
+	  , isRegExp    = __webpack_require__(127)
+	  , getFlags    = __webpack_require__(185)
+	  , RegExpProto = RegExp.prototype;
+
+	var $RegExpStringIterator = function(regexp, string){
+	  this._r = regexp;
+	  this._s = string;
+	};
+
+	__webpack_require__(135)($RegExpStringIterator, 'RegExp String', function next(){
+	  var match = this._r.exec(this._s);
+	  return {value: match, done: match === null};
+	});
+
+	$export($export.P, 'String', {
+	  matchAll: function matchAll(regexp){
+	    defined(this);
+	    if(!isRegExp(regexp))throw TypeError(regexp + ' is not a regexp!');
+	    var S     = String(this)
+	      , flags = 'flags' in RegExpProto ? String(regexp.flags) : getFlags.call(regexp)
+	      , rx    = new RegExp(regexp.source, ~flags.indexOf('g') ? flags : 'g' + flags);
+	    rx.lastIndex = toLength(regexp.lastIndex);
+	    return new $RegExpStringIterator(rx, S);
+	  }
+	});
+
+/***/ },
+/* 250 */
+/***/ function(module, exports, __webpack_require__) {
+
+	// https://github.com/tc39/proposal-object-getownpropertydescriptors
 	var $export    = __webpack_require__(7)
 	  , ownKeys    = __webpack_require__(217)
 	  , toIObject  = __webpack_require__(27)
@@ -6250,12 +6325,12 @@
 	});
 
 /***/ },
-/* 248 */
+/* 251 */
 /***/ function(module, exports, __webpack_require__) {
 
-	// http://goo.gl/XkBrjD
+	// https://github.com/tc39/proposal-object-values-entries
 	var $export = __webpack_require__(7)
-	  , $values = __webpack_require__(249)(false);
+	  , $values = __webpack_require__(252)(false);
 
 	$export($export.S, 'Object', {
 	  values: function values(it){
@@ -6264,7 +6339,7 @@
 	});
 
 /***/ },
-/* 249 */
+/* 252 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var getKeys   = __webpack_require__(25)
@@ -6285,12 +6360,12 @@
 	};
 
 /***/ },
-/* 250 */
+/* 253 */
 /***/ function(module, exports, __webpack_require__) {
 
-	// http://goo.gl/XkBrjD
+	// https://github.com/tc39/proposal-object-values-entries
 	var $export  = __webpack_require__(7)
-	  , $entries = __webpack_require__(249)(true);
+	  , $entries = __webpack_require__(252)(true);
 
 	$export($export.S, 'Object', {
 	  entries: function entries(it){
@@ -6299,21 +6374,113 @@
 	});
 
 /***/ },
-/* 251 */
+/* 254 */
+/***/ function(module, exports, __webpack_require__) {
+
+	'use strict';
+	var $export         = __webpack_require__(7)
+	  , toObject        = __webpack_require__(54)
+	  , aFunction       = __webpack_require__(19)
+	  , $defineProperty = __webpack_require__(9);
+
+	// B.2.2.2 Object.prototype.__defineGetter__(P, getter)
+	__webpack_require__(5) && $export($export.P + __webpack_require__(255), 'Object', {
+	  __defineGetter__: function __defineGetter__(P, getter){
+	    $defineProperty.f(toObject(this), P, {get: aFunction(getter), enumerable: true, configurable: true});
+	  }
+	});
+
+/***/ },
+/* 255 */
+/***/ function(module, exports, __webpack_require__) {
+
+	// Forced replacement prototype accessors methods
+	module.exports = __webpack_require__(47)|| !__webpack_require__(6)(function(){
+	  var K = Math.random();
+	  // In FF throws only define methods
+	  __defineSetter__.call(null, K, function(){ /* empty */});
+	  delete __webpack_require__(2)[K];
+	});
+
+/***/ },
+/* 256 */
+/***/ function(module, exports, __webpack_require__) {
+
+	'use strict';
+	var $export         = __webpack_require__(7)
+	  , toObject        = __webpack_require__(54)
+	  , aFunction       = __webpack_require__(19)
+	  , $defineProperty = __webpack_require__(9);
+
+	// B.2.2.3 Object.prototype.__defineSetter__(P, setter)
+	__webpack_require__(5) && $export($export.P + __webpack_require__(255), 'Object', {
+	  __defineSetter__: function __defineSetter__(P, setter){
+	    $defineProperty.f(toObject(this), P, {set: aFunction(setter), enumerable: true, configurable: true});
+	  }
+	});
+
+/***/ },
+/* 257 */
+/***/ function(module, exports, __webpack_require__) {
+
+	'use strict';
+	var $export                  = __webpack_require__(7)
+	  , toObject                 = __webpack_require__(54)
+	  , toPrimitive              = __webpack_require__(14)
+	  , getPrototypeOf           = __webpack_require__(55)
+	  , getOwnPropertyDescriptor = __webpack_require__(46).f;
+
+	// B.2.2.4 Object.prototype.__lookupGetter__(P)
+	__webpack_require__(5) && $export($export.P + __webpack_require__(255), 'Object', {
+	  __lookupGetter__: function __lookupGetter__(P){
+	    var O = toObject(this)
+	      , K = toPrimitive(P, true)
+	      , D;
+	    do {
+	      if(D = getOwnPropertyDescriptor(O, K))return D.get;
+	    } while(O = getPrototypeOf(O));
+	  }
+	});
+
+/***/ },
+/* 258 */
+/***/ function(module, exports, __webpack_require__) {
+
+	'use strict';
+	var $export                  = __webpack_require__(7)
+	  , toObject                 = __webpack_require__(54)
+	  , toPrimitive              = __webpack_require__(14)
+	  , getPrototypeOf           = __webpack_require__(55)
+	  , getOwnPropertyDescriptor = __webpack_require__(46).f;
+
+	// B.2.2.5 Object.prototype.__lookupSetter__(P)
+	__webpack_require__(5) && $export($export.P + __webpack_require__(255), 'Object', {
+	  __lookupSetter__: function __lookupSetter__(P){
+	    var O = toObject(this)
+	      , K = toPrimitive(P, true)
+	      , D;
+	    do {
+	      if(D = getOwnPropertyDescriptor(O, K))return D.set;
+	    } while(O = getPrototypeOf(O));
+	  }
+	});
+
+/***/ },
+/* 259 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/DavidBruant/Map-Set.prototype.toJSON
 	var $export  = __webpack_require__(7);
 
-	$export($export.P + $export.R, 'Map', {toJSON: __webpack_require__(252)('Map')});
+	$export($export.P + $export.R, 'Map', {toJSON: __webpack_require__(260)('Map')});
 
 /***/ },
-/* 252 */
+/* 260 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/DavidBruant/Map-Set.prototype.toJSON
 	var classof = __webpack_require__(71)
-	  , from    = __webpack_require__(253);
+	  , from    = __webpack_require__(261);
 	module.exports = function(NAME){
 	  return function toJSON(){
 	    if(classof(this) != NAME)throw TypeError(NAME + "#toJSON isn't generic");
@@ -6322,7 +6489,7 @@
 	};
 
 /***/ },
-/* 253 */
+/* 261 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var forOf = __webpack_require__(194);
@@ -6335,16 +6502,16 @@
 
 
 /***/ },
-/* 254 */
+/* 262 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/DavidBruant/Map-Set.prototype.toJSON
 	var $export  = __webpack_require__(7);
 
-	$export($export.P + $export.R, 'Set', {toJSON: __webpack_require__(252)('Set')});
+	$export($export.P + $export.R, 'Set', {toJSON: __webpack_require__(260)('Set')});
 
 /***/ },
-/* 255 */
+/* 263 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/ljharb/proposal-global
@@ -6353,7 +6520,7 @@
 	$export($export.S, 'System', {global: __webpack_require__(2)});
 
 /***/ },
-/* 256 */
+/* 264 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://github.com/ljharb/proposal-is-error
@@ -6367,7 +6534,7 @@
 	});
 
 /***/ },
-/* 257 */
+/* 265 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://gist.github.com/BrendanEich/4294d5c212a6d2254703
@@ -6383,7 +6550,7 @@
 	});
 
 /***/ },
-/* 258 */
+/* 266 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://gist.github.com/BrendanEich/4294d5c212a6d2254703
@@ -6399,7 +6566,7 @@
 	});
 
 /***/ },
-/* 259 */
+/* 267 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://gist.github.com/BrendanEich/4294d5c212a6d2254703
@@ -6420,7 +6587,7 @@
 	});
 
 /***/ },
-/* 260 */
+/* 268 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// https://gist.github.com/BrendanEich/4294d5c212a6d2254703
@@ -6441,10 +6608,10 @@
 	});
 
 /***/ },
-/* 261 */
+/* 269 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata                  = __webpack_require__(262)
+	var metadata                  = __webpack_require__(270)
 	  , anObject                  = __webpack_require__(10)
 	  , toMetaKey                 = metadata.key
 	  , ordinaryDefineOwnMetadata = metadata.set;
@@ -6454,7 +6621,7 @@
 	}});
 
 /***/ },
-/* 262 */
+/* 270 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var Map     = __webpack_require__(199)
@@ -6510,10 +6677,10 @@
 	};
 
 /***/ },
-/* 263 */
+/* 271 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata               = __webpack_require__(262)
+	var metadata               = __webpack_require__(270)
 	  , anObject               = __webpack_require__(10)
 	  , toMetaKey              = metadata.key
 	  , getOrCreateMetadataMap = metadata.map
@@ -6530,10 +6697,10 @@
 	}});
 
 /***/ },
-/* 264 */
+/* 272 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata               = __webpack_require__(262)
+	var metadata               = __webpack_require__(270)
 	  , anObject               = __webpack_require__(10)
 	  , getPrototypeOf         = __webpack_require__(55)
 	  , ordinaryHasOwnMetadata = metadata.has
@@ -6552,12 +6719,12 @@
 	}});
 
 /***/ },
-/* 265 */
+/* 273 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var Set                     = __webpack_require__(202)
-	  , from                    = __webpack_require__(253)
-	  , metadata                = __webpack_require__(262)
+	  , from                    = __webpack_require__(261)
+	  , metadata                = __webpack_require__(270)
 	  , anObject                = __webpack_require__(10)
 	  , getPrototypeOf          = __webpack_require__(55)
 	  , ordinaryOwnMetadataKeys = metadata.keys
@@ -6576,10 +6743,10 @@
 	}});
 
 /***/ },
-/* 266 */
+/* 274 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata               = __webpack_require__(262)
+	var metadata               = __webpack_require__(270)
 	  , anObject               = __webpack_require__(10)
 	  , ordinaryGetOwnMetadata = metadata.get
 	  , toMetaKey              = metadata.key;
@@ -6590,10 +6757,10 @@
 	}});
 
 /***/ },
-/* 267 */
+/* 275 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata                = __webpack_require__(262)
+	var metadata                = __webpack_require__(270)
 	  , anObject                = __webpack_require__(10)
 	  , ordinaryOwnMetadataKeys = metadata.keys
 	  , toMetaKey               = metadata.key;
@@ -6603,10 +6770,10 @@
 	}});
 
 /***/ },
-/* 268 */
+/* 276 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata               = __webpack_require__(262)
+	var metadata               = __webpack_require__(270)
 	  , anObject               = __webpack_require__(10)
 	  , getPrototypeOf         = __webpack_require__(55)
 	  , ordinaryHasOwnMetadata = metadata.has
@@ -6624,10 +6791,10 @@
 	}});
 
 /***/ },
-/* 269 */
+/* 277 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata               = __webpack_require__(262)
+	var metadata               = __webpack_require__(270)
 	  , anObject               = __webpack_require__(10)
 	  , ordinaryHasOwnMetadata = metadata.has
 	  , toMetaKey              = metadata.key;
@@ -6638,10 +6805,10 @@
 	}});
 
 /***/ },
-/* 270 */
+/* 278 */
 /***/ function(module, exports, __webpack_require__) {
 
-	var metadata                  = __webpack_require__(262)
+	var metadata                  = __webpack_require__(270)
 	  , anObject                  = __webpack_require__(10)
 	  , aFunction                 = __webpack_require__(19)
 	  , toMetaKey                 = metadata.key
@@ -6658,7 +6825,7 @@
 	}});
 
 /***/ },
-/* 271 */
+/* 279 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var $export = __webpack_require__(7)
@@ -6669,7 +6836,7 @@
 	});
 
 /***/ },
-/* 272 */
+/* 280 */
 /***/ function(module, exports, __webpack_require__) {
 
 	var $iterators    = __webpack_require__(180)
@@ -6696,14 +6863,14 @@
 	}
 
 /***/ },
-/* 273 */
+/* 281 */
 /***/ function(module, exports, __webpack_require__) {
 
 	// ie9- setTimeout & setInterval additional parameters fix
 	var global     = __webpack_require__(2)
 	  , $export    = __webpack_require__(7)
 	  , invoke     = __webpack_require__(74)
-	  , partial    = __webpack_require__(274)
+	  , partial    = __webpack_require__(282)
 	  , navigator  = global.navigator
 	  , MSIE       = !!navigator && /MSIE .\./.test(navigator.userAgent); // <- dirty ie9- check
 	var wrap = function(set){
@@ -6721,11 +6888,11 @@
 	});
 
 /***/ },
-/* 274 */
+/* 282 */
 /***/ function(module, exports, __webpack_require__) {
 
 	'use strict';
-	var path      = __webpack_require__(275)
+	var path      = __webpack_require__(283)
 	  , invoke    = __webpack_require__(74)
 	  , aFunction = __webpack_require__(19);
 	module.exports = function(/* ...pargs */){
@@ -6749,7 +6916,7 @@
 	};
 
 /***/ },
-/* 275 */
+/* 283 */
 /***/ function(module, exports, __webpack_require__) {
 
 	module.exports = __webpack_require__(2);