+++ modules/_core.js
@@ -1,2 +1,2 @@
-var core = module.exports = {version: '2.1.5'};
+var core = module.exports = {version: '2.2.0'};
 if(typeof __e == 'number')__e = core; // eslint-disable-line no-undef